+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,7 +3,7 @@ package kitchenpos;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.order.domain.*;
 
 import java.math.BigDecimal;
